% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_plot.R
\name{dgev_torch}
\alias{dgev_torch}
\title{Quantile function for d-GEV using Torch}
\usage{
dgev_torch(p, mu, sigma, xi)
}
\arguments{
\item{p}{Probability}

\item{mu}{Location parameter}

\item{sigma}{Scale parameter}

\item{xi}{Shape parameter}
}
\description{
Quantile function for d-GEV using Torch
}
