% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgevbhm-package.R
\docType{package}
\name{dgevbhm-package}
\alias{dgevbhm}
\alias{dgevbhm-package}
\title{The 'dgevbhm' package.}
\description{
Bayesian Hierarchical Modeling for Duration-Dependent GEV
}
\author{
\strong{Maintainer}: Alexander Lee Rischmuller \email{alex.rischmuller@gmail.com}

Authors:
\itemize{
  \item Alexander Lee Rischmuller \email{alex.rischmuller@gmail.com}
}

}
\keyword{internal}
