% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit_xi_j}
\alias{fit_xi_j}
\title{Result of xi_d Model Fit}
\format{
An object of class \code{list} of length 8.
}
\usage{
fit_xi_j
}
\description{
Pre-computed BHM fit using the shape parameter hierarchical with respect to space.
}
\keyword{datasets}
