% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PP_plot.R
\name{gev_cdf_torch}
\alias{gev_cdf_torch}
\title{GEV Cumulative Distribution Function using Torch}
\usage{
gev_cdf_torch(x, mu, sigma, xi)
}
\arguments{
\item{x}{Tensor. The data points.}

\item{mu}{Tensor. Location parameter.}

\item{sigma}{Tensor. Scale parameter.}

\item{xi}{Tensor. Shape parameter.}
}
\description{
Calculates the CDF for the Generalized Extreme Value distribution using
tensor operations.
}
