% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_plot.R
\name{gev_mat}
\alias{gev_mat}
\title{Quantile function for d-GEV using Base R}
\usage{
gev_mat(p, mu_mat, sigma_mat, xi_vec, xi_dim)
}
\arguments{
\item{p}{Probability (scalar)}

\item{mu_mat}{Location parameter matrix (Durations x Samples)}

\item{sigma_mat}{Scale parameter matrix (Durations x Samples)}

\item{xi_vec}{Shape parameter vector}

\item{xi_dim}{Character. "duration" for shp_d='d', "sample" for shp_d='j'}
}
\description{
Quantile function for d-GEV using Base R
}
