% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_plot.R
\name{idf_plot}
\alias{idf_plot}
\title{Plot IDF Curves using Base R}
\usage{
idf_plot(fit, j = NULL, rp, alpha = 0.05)
}
\arguments{
\item{fit}{List. The output object from the dgev_bhm function.}

\item{j}{Integer. Station index.}

\item{rp}{Numeric vector. Return periods (max 3).}

\item{alpha}{Numeric. Significance level.}
}
\description{
Plot IDF Curves using Base R
}
