% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_plot.R
\name{idf_plot_torch}
\alias{idf_plot_torch}
\title{Plot IDF Curves using Torch}
\usage{
idf_plot_torch(fit, j, rp, alpha = 0.05, cores = 4)
}
\arguments{
\item{fit}{List. The output object from the dgev_bhm function.}

\item{j}{Integer. Station index.}

\item{rp}{Numeric vector. Return periods (max 3).}

\item{alpha}{Numeric. Significance level.}

\item{cores}{Integer. Threads for torch.}
}
\description{
Plot IDF Curves using Torch
}
