% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_pred.R
\name{post_pred_plot}
\alias{post_pred_plot}
\title{Posterior Predictive Histogram Plot (Base R)}
\usage{
post_pred_plot(fit, stat, d, j)
}
\arguments{
\item{fit}{List. Output from \code{dgev_bhm}.}

\item{stat}{Numeric (0-1) or Character ("median", "max", "min"). The statistic to check.}

\item{d}{Integer. Index of the duration to plot.}

\item{j}{Integer. Index of the station to plot.}
}
\description{
Generates a histogram of the posterior predictive distribution for a specific
statistic and compares it against the observed (empirical) value.
}
