% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_pred.R
\name{post_pred_pval}
\alias{post_pred_pval}
\title{Posterior Predictive P-value (PPP)}
\usage{
post_pred_pval(fit, stat, d, j)
}
\arguments{
\item{fit}{List. Output from \code{dgev_bhm}.}

\item{stat}{Numeric (0-1) or Character ("median", "max", "min"). The statistic to check.}

\item{d}{Integer. Index of the duration.}

\item{j}{Integer. Index of the station.}
}
\value{
Numeric. The calculated PPP.
}
\description{
Calculates the posterior predictive p-value for a given statistic.
A value near 0.5 indicates a good model fit.
}
