% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_diag.R
\name{pp_plot_torch}
\alias{pp_plot_torch}
\title{Posterior Predictive Histogram Plot}
\usage{
pp_plot_torch(fit, stat, d, j, cores = 4)
}
\arguments{
\item{fit}{List. Output from \code{dgev_bhm}.}

\item{stat}{Numeric (0-1) or Character ("median", "max", "min"). The statistic to check.}

\item{d}{Integer. Index of the duration to plot.}

\item{j}{Integer. Index of the station to plot.}

\item{cores}{Integer. Number of threads for torch.}
}
\description{
Generates a histogram of the posterior predictive distribution for a specific
statistic and compares it against the observed (empirical) value.
}
