% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_diag.R
\name{ppp_torch}
\alias{ppp_torch}
\title{Posterior Predictive P-value (PPP)}
\usage{
ppp_torch(fit, stat, d, j, cores = 4)
}
\arguments{
\item{fit}{List. Output from \code{dgev_bhm}.}

\item{stat}{Numeric (0-1) or Character ("median", "max", "min"). The statistic to check.}

\item{d}{Integer. Index of the duration.}

\item{j}{Integer. Index of the station.}

\item{cores}{Integer. Number of threads for torch.}
}
\value{
Numeric. The calculated PPP value.
}
\description{
Calculates the posterior predictive p-value for a given statistic.
A value near 0.5 indicates a good model fit.
}
