% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_level.R
\name{return_level}
\alias{return_level}
\title{Plot Return Level Curves for d-GEV}
\usage{
return_level(fit, max_rp = 100, j, d, alpha = 0.05)
}
\arguments{
\item{fit}{List. Output from \code{dgev_bhm}.}

\item{max_rp}{Numeric or Character. Maximum return period to plot. Use "data"
to cap at the record length.}

\item{j}{Integer. Index of the station.}

\item{d}{Integer. Index of the duration.}

\item{alpha}{Numeric. Significance level for the credible interval (default 0.05).}
}
\description{
Calculates and plots the relationship between return periods and precipitation
intensity (return levels) with Bayesian credible intervals.
}
