% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UK_data_full}
\alias{UK_data_full}
\title{UK Extreme Rainfall Data}
\format{
A list with elements:
\describe{
\item{y}{A list of matrices duration dependent annual maxima for each station. year x duration}
\item{amp}{Annual mean precipitation for each station.}
\item{latlon}{Latitude and longitude for each station.}
\item{stationID}{station ID for each station.}
\item{durs}{Durations corresponding to the columns of the y matrices.}
}
}
\usage{
UK_data_full
}
\description{
A dataset containing annual maxima for 47 stations.
The full list of processed data.
}
\keyword{datasets}
