% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{waic}
\alias{waic}
\title{Calculate WAIC for d-GEV BHM}
\usage{
waic(fit, cores = 4)
}
\arguments{
\item{fit}{List. Output from \code{dgev_bhm}.}

\item{cores}{Integer. Number of threads for torch (default 4).}
}
\value{
Numeric. The calculated WAIC value.
}
\description{
Computes the Watanabe-Akaike Information Criterion (WAIC) for the fitted
Bayesian Hierarchical Model.
}
