% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgev_bhm.R
\name{dgev_bhm}
\alias{dgev_bhm}
\title{Bayesian Hierarchical Model for d-GEV}
\usage{
dgev_bhm(data, iter = 2000, cores = 4, shp_d = NULL)
}
\arguments{
\item{data}{A list containing processed station data (output from \code{data_read}).}

\item{iter}{Integer. Number of iterations for MCMC sampling.}

\item{cores}{Integer. Number of cores for parallel sampling.}

\item{shp_d}{Character. Dimension for shape parameter: "d" for durational, "j" for spatial.}
}
\value{
A list containing MCMC samples (pars), original data, and metadata.
}
\description{
Fits a Bayesian Hierarchical Model using Stan for duration-dependent
Generalized Extreme Value distributions.
}
