% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{gev_lpdf_robust}
\alias{gev_lpdf_robust}
\title{Calculate WAIC for d-GEV BHM}
\usage{
gev_lpdf_robust(x, mu, sigma, xi)
}
\arguments{
\item{fit}{List. Output from \code{dgev_bhm}.}

\item{cores}{Integer. Number of threads for torch (default 4).}
}
\value{
Numeric. The calculated WAIC value.
}
\description{
Computes the Watanabe-Akaike Information Criterion (WAIC) for the fitted
Bayesian Hierarchical Model.
}
